pyproject.toml
[project]
name = "winter2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "markdown-it-py>=4.0.0",
]
